Updated BG Layer 2 files
It was brought to my attention that the contents of my layer 2s were
just a smidge too tall and it was blocking the layer1s, so I resized
them. They're still the same height, it's just cropped off about 60px on
the bottom and then that 60px was added back onto the top of the image,
so it's literally the same size.