Title and Game Over Screens
The title is pretty self-explanatory. The Game Over screen has a
"pressed" button version (the white button) and an "unpressed" button
version which is just the waiting screen until the player takes action.
As for the title screen, if you can flip flop between the two images, it
will look like the text is blinking. I hope that makes sense.